+++ app/settings/page.tsx
@@ -1,5 +1,16 @@
 import AuthGuard from "@/app/components/auth-guard";
 import UserSettings from "../components/user-settings";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "계정 설정",
+  description:
+    "Oneiri 계정 설정을 관리하세요. 프로필 정보 수정, 개인정보 보호 설정 등을 변경할 수 있습니다.",
+  robots: {
+    index: false,
+    follow: true,
+  },
+};
 
 export default function SettingsPage() {
   return (
