+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "my-v0-project",
-  "version": "0.1.0",
+  "name": "oneiri",
+  "version": "1.1.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -80,6 +80,8 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "conventional-changelog-conventionalcommits": "^9.0.0",
+    "eslint": "^9.29.0",
+    "eslint-config-next": "15.3.3",
     "postcss": "^8.5",
     "semantic-release": "^24.2.5",
     "tailwindcss": "^3.4.17",
