+++ pnpm-workspace.yaml
@@ -1,2 +1,3 @@
 ignoredBuiltDependencies:
   - '@vercel/speed-insights'
+  - unrs-resolver
